fix(project): currency int to double issue

+++ lib/features/cart/presentation/widgets/footer/cart_footer.dart
@@ -13,7 +13,7 @@ class CartFooter extends StatelessWidget {
   });
 
   final Function() onSaved;
-  final ValueChanged<int> onCompleted;
+  final ValueChanged<double> onCompleted;
 
   @override
   Widget build(BuildContext context) {
@@ -55,7 +55,8 @@ class CartFooter extends StatelessWidget {
                         // TODO: handle int string decimal
                         if (cartState is CartDetailLoadSuccess) {
                           onCompleted(
-                              int.parse(cartState.previewOrderPrice.total));
+                            double.parse(cartState.previewOrderPrice.total),
+                          );
                         }
                       },
                 child: cartState is CartDetailActionInProgress
